refactored tranches hooks
