+++ src/components/pages/Tranches/TranchesPage.tsx
@@ -14,49 +14,27 @@ import {
 import { useTranslation } from "react-i18next";
 import { MdSwapHoriz } from "react-icons/md";
 import CopyrightSpacer from "../../shared/CopyrightSpacer";
-import { useQuery } from "react-query";
-
-import ERC20ABI from "../../../rari-sdk/abi/ERC20.json";
-import {
-  smallStringUsdFormatter,
-  smallUsdFormatter,
-} from "../../../utils/bigUtils";
 import DepositModal from "./SaffronDepositModal";
-import { SaffronProvider, useSaffronContracts } from "./SaffronContext";
+import { SaffronProvider } from "./SaffronContext";
 import { SimpleTooltip } from "../../shared/SimpleTooltip";
 import { WarningTwoIcon } from "@chakra-ui/icons";
 
 
 // Hooks
 import { useIsSmallScreen } from "hooks/useIsSmallScreen";
-import { useSaffronData } from 'hooks/tranches/useSaffronData'
+import { 
+  TranchePool, 
+  TrancheRating, 
+  tranchePoolIndex,
+  trancheRatingIndex,
+  useEpochEndDate, 
+  useEstimatedSFI, 
+  usePrincipal, 
+  usePrincipalBalance, 
+  useSaffronData } from 'hooks/tranches/useSaffronData'
 import { useSFIDistributions } from "hooks/tranches/useSFIDistributions";
 import { useSFIEarnings } from "hooks/tranches/useSFIEarnings";
 import { useAuthedCallback } from "hooks/useAuthedCallback";
-
-export enum TranchePool {
-  DAI = "DAI",
-  USDC = "USDC",
-}
-
-export enum TrancheRating {
-  S = "S",
-  AA = "AA",
-  A = "A",
-}
-
-export const trancheRatingIndex = (trancheRating: TrancheRating) => {
-  return trancheRating === TrancheRating.S
-    ? 0
-    : trancheRating === TrancheRating.AA
-      ? 1
-      : 2;
-};
-
-export const tranchePoolIndex = (tranchePool: TranchePool) => {
-  // TODO: CHANGE USDC TO WHATEVER IT BECOMES LATER
-  return tranchePool === TranchePool.DAI ? 9 : 0;
-};
 
 const WrappedTranchePage = React.memo(() => {
   return (
@@ -70,9 +48,7 @@ export default WrappedTranchePage;
 
 const TranchePage = () => {
   const { isAuthed } = useRari();
-
   const { t } = useTranslation();
-
   const isMobile = useIsSmallScreen();
 
   return (
@@ -191,7 +167,6 @@ const TranchePage = () => {
 
 export const TranchesRatingInfo = () => {
   const { t } = useTranslation();
-
   const isMobile = useIsSmallScreen();
 
   return (
@@ -237,11 +212,7 @@ export const TrancheRatingColumn = ({
   trancheRating: TrancheRating;
 }) => {
   const { t } = useTranslation();
-
   const isMobile = useIsSmallScreen();
-
-  const { saffronPool } = useSaffronData();
-
   const data = useSFIEarnings()
 
   return (
@@ -344,31 +315,9 @@ export const TrancheColumn = ({
   const { t } = useTranslation();
   const isMobile = useIsSmallScreen();
 
-  const { rari, address } = useRari();
-  const { saffronData, saffronPool, fetchCurrentEpoch } = useSaffronData();
+  const { saffronData } = useSaffronData();
 
-  const { data: principal } = useQuery(
-    tranchePool + trancheRating + " principal " + address,
-    async () => {
-      //TODO: ADD USDC POOL
-
-      const currentEpoch = await fetchCurrentEpoch();
-
-      const tranchePToken = new rari.web3.eth.Contract(
-        ERC20ABI as any,
-        await saffronPool.methods
-          .principal_token_addresses(
-            currentEpoch,
-            trancheRatingIndex(trancheRating)
-          )
-          .call()
-      );
-
-      return smallUsdFormatter(
-        parseInt(await tranchePToken.methods.balanceOf(address).call()) / 1e18
-      ).replace("$", "");
-    }
-  );
+  const principal = usePrincipal(tranchePool, trancheRating)
 
   const {
     isOpen: isDepositModalOpen,
@@ -445,18 +394,7 @@ export const TrancheColumn = ({
 
 export const RedemptionDate = () => {
   const { t } = useTranslation();
-
-  const { saffronPool, fetchCurrentEpoch } = useSaffronData();
-
-  const { data } = useQuery("epochEndDate", async () => {
-    const currentEpoch = await fetchCurrentEpoch();
-
-    const endDate = new Date(
-      (await saffronPool.methods.get_epoch_end(currentEpoch).call()) * 1000
-    );
-
-    return { currentEpoch, endDate };
-  });
+  const { data } =  useEpochEndDate()
 
   return (
     <Column expand mainAxisAlignment="center" crossAxisAlignment="center">
@@ -478,108 +416,8 @@ export const RedemptionDate = () => {
 export const PrincipalAmount = () => {
   const { t } = useTranslation();
 
-  const { rari, address } = useRari();
-  const { saffronPool, saffronStrategy, fetchCurrentEpoch } = useSaffronData();
-
-  const { data: principal } = useQuery(
-    "principalBalance " + address,
-    async () => {
-      const currentEpoch = await fetchCurrentEpoch();
-
-      const sTranchePToken = new rari.web3.eth.Contract(
-        ERC20ABI as any,
-        await saffronPool.methods
-          .principal_token_addresses(currentEpoch, 0)
-          .call()
-      );
-
-      const aTranchePToken = new rari.web3.eth.Contract(
-        ERC20ABI as any,
-        await saffronPool.methods
-          .principal_token_addresses(currentEpoch, 2)
-          .call()
-      );
-
-      return smallStringUsdFormatter(
-        rari.web3.utils.fromWei(
-          rari.web3.utils
-            .toBN(await sTranchePToken.methods.balanceOf(address).call())
-            .add(
-              rari.web3.utils.toBN(
-                await aTranchePToken.methods.balanceOf(address).call()
-              )
-            )
-        )
-      );
-    }
-  );
-
-  const { data: estimatedSFI } = useQuery(
-    "estimatedSFI " + address,
-    async () => {
-      // TODO: ADD USDC
-
-      const DAI_SFI_REWARDS =
-        parseInt(
-          await saffronStrategy.methods
-            .pool_SFI_rewards(tranchePoolIndex(TranchePool.DAI))
-            .call()
-        ) / 1e18;
-
-      const SFI_multipliers = await saffronPool.methods
-        .TRANCHE_SFI_MULTIPLIER()
-        .call();
-
-      const currentEpoch = await fetchCurrentEpoch();
-
-      const dsecSToken = new rari.web3.eth.Contract(
-        ERC20ABI as any,
-        await saffronPool.methods
-          .principal_token_addresses(
-            currentEpoch,
-            trancheRatingIndex(TrancheRating.S)
-          )
-          .call()
-      );
-
-      // TODO ADD AA POOL
-
-      const dsecSSupply = await dsecSToken.methods.totalSupply().call();
-
-      const sPoolSFIEarned =
-        DAI_SFI_REWARDS *
-        (SFI_multipliers[trancheRatingIndex(TrancheRating.S)] / 100000) *
-        // If supply is zero we will get NaN for dividing by zero
-        (dsecSSupply === "0"
-          ? 0
-          : (await dsecSToken.methods.balanceOf(address).call()) / dsecSSupply);
-
-      const dsecAToken = new rari.web3.eth.Contract(
-        ERC20ABI as any,
-        await saffronPool.methods
-          .principal_token_addresses(
-            currentEpoch,
-            trancheRatingIndex(TrancheRating.A)
-          )
-          .call()
-      );
-
-      const dsecASupply = await dsecAToken.methods.totalSupply().call();
-
-      const aPoolSFIEarned =
-        DAI_SFI_REWARDS *
-        (SFI_multipliers[trancheRatingIndex(TrancheRating.A)] / 100000) *
-        // If supply is zero we will get NaN for dividing by zero
-        (dsecASupply === "0"
-          ? 0
-          : (await dsecAToken.methods.balanceOf(address).call()) / dsecASupply);
-
-      return (
-        smallUsdFormatter(sPoolSFIEarned + aPoolSFIEarned).replace("$", "") +
-        " SFI"
-      );
-    }
-  );
+  const principal = usePrincipalBalance()
+  const estimatedSFI = useEstimatedSFI()
 
   return (
     <Column expand mainAxisAlignment="center" crossAxisAlignment="center">
@@ -625,7 +463,7 @@ export const SFIDistributions = () => {
   const { rari } = useRari();
 
   const sfiDistributions = useSFIDistributions()
- 
+
 
   // TODO: ADD USDC
   return (
