+++ src/hooks/tranches/useSFIDistributions.ts
@@ -3,21 +3,14 @@ import { useQuery } from "react-query";
 import { useRari } from "context/RariContext";
 import { useSaffronData } from "./useSaffronData";
 
-export enum TranchePool {
-    DAI = "DAI",
-    USDC = "USDC",
-  }
 
-export const tranchePoolIndex = (tranchePool: TranchePool) => {
-    // TODO: CHANGE USDC TO WHATEVER IT BECOMES LATER
-    return tranchePool === TranchePool.DAI ? 9 : 0;
-  };
+import { tranchePoolIndex, TranchePool } from './useSaffronData'
+
 
 export const useSFIDistributions = () => {
 
     const { rari } = useRari()
     const { saffronStrategy } = useSaffronData()
-
 
 
     const { data: sfiDistributions } = useQuery(
